Added string concat bytecode instructions to avoid doing concatenation in go code (#1218)
* Reworked CONCAT to be variable arguments at AST level
* Added a better simplify logic into CONCAT to simplify literals that are next to each other and to remove empty literals
* Removed apply stuff as it was only designed to make CONCAT working
* Added support for CONCAT into bytecode, including string allocation and append functionality

+++ expr/partiql/y.go
@@ -231,7 +231,7 @@ const yyErrCode = 2
 const yyInitialStackSize = 16
 
 //line yacctab:1
-var yyExca = [...]int{
+var yyExca = [...]int16{
 	-1, 1,
 	1, -1,
 	-2, 0,
@@ -253,7 +253,7 @@ const yyPrivate = 57344
 
 const yyLast = 1618
 
-var yyAct = [...]int{
+var yyAct = [...]int16{
 	15, 285, 289, 278, 170, 270, 228, 256, 178, 13,
 	110, 97, 14, 9, 17, 193, 126, 125, 172, 36,
 	134, 157, 100, 209, 7, 88, 11, 60, 61, 62,
@@ -418,7 +418,7 @@ var yyAct = [...]int{
 	77, 58, 59, 60, 61, 62, 63, 64,
 }
 
-var yyPact = [...]int{
+var yyPact = [...]int16{
 	236, -1000, 263, 119, 151, 233, 151, 230, -1000, 412,
 	-1000, 229, 173, 129, -1000, 680, -1000, -1000, 172, 171,
 	170, 169, 168, 163, 160, 157, 156, -46, 155, 154,
@@ -452,14 +452,14 @@ var yyPact = [...]int{
 	-1000, 223, -1000, -1000, -1000,
 }
 
-var yyPgo = [...]int{
+var yyPgo = [...]int16{
 	0, 292, 0, 289, 14, 148, 288, 8, 7, 287,
 	286, 283, 279, 11, 278, 277, 13, 19, 4, 43,
 	6, 9, 12, 10, 275, 21, 2, 5, 274, 273,
 	3, 1, 272, 271,
 }
 
-var yyR1 = [...]int{
+var yyR1 = [...]int8{
 	0, 1, 19, 6, 6, 11, 11, 12, 12, 22,
 	22, 22, 22, 5, 3, 3, 3, 3, 3, 3,
 	3, 3, 4, 4, 10, 10, 16, 16, 2, 2,
@@ -476,7 +476,7 @@ var yyR1 = [...]int{
 	30, 31, 31,
 }
 
-var yyR2 = [...]int{
+var yyR2 = [...]int8{
 	0, 12, 10, 2, 0, 1, 0, 6, 7, 3,
 	2, 1, 1, 2, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 3, 1, 1, 1, 0, 1, 4,
@@ -493,7 +493,7 @@ var yyR2 = [...]int{
 	2, 0, 2,
 }
 
-var yyChk = [...]int{
+var yyChk = [...]int16{
 	-1000, -1, -11, -12, 16, 7, 55, -17, 53, -16,
 	18, -17, 20, -21, -22, -2, 87, -4, 28, 31,
 	29, 30, 32, 43, 44, 37, 38, 70, 33, 34,
@@ -527,7 +527,7 @@ var yyChk = [...]int{
 	-15, 22, -31, 23, 24,
 }
 
-var yyDef = [...]int{
+var yyDef = [...]int16{
 	6, -2, 0, 5, 0, 27, 0, 0, 107, 0,
 	26, 0, 0, 4, 83, 11, 12, 28, 0, 0,
 	0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
@@ -561,7 +561,7 @@ var yyDef = [...]int{
 	124, 0, 2, 119, 120,
 }
 
-var yyTok1 = [...]int{
+var yyTok1 = [...]int8{
 	1, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
@@ -577,7 +577,7 @@ var yyTok1 = [...]int{
 	3, 3, 3, 59, 3, 60,
 }
 
-var yyTok2 = [...]int{
+var yyTok2 = [...]int8{
 	2, 3, 4, 5, 6, 7, 8, 9, 10, 11,
 	12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
 	22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
@@ -589,7 +589,7 @@ var yyTok2 = [...]int{
 	96,
 }
 
-var yyTok3 = [...]int{
+var yyTok3 = [...]int8{
 	0,
 }
 
@@ -671,9 +671,9 @@ func yyErrorMessage(state, lookAhead int) string {
 	expected := make([]int, 0, 4)
 
 	// Look for shiftable tokens.
-	base := yyPact[state]
+	base := int(yyPact[state])
 	for tok := TOKSTART; tok-1 < len(yyToknames); tok++ {
-		if n := base + tok; n >= 0 && n < yyLast && yyChk[yyAct[n]] == tok {
+		if n := base + tok; n >= 0 && n < yyLast && int(yyChk[int(yyAct[n])]) == tok {
 			if len(expected) == cap(expected) {
 				return res
 			}
@@ -683,13 +683,13 @@ func yyErrorMessage(state, lookAhead int) string {
 
 	if yyDef[state] == -2 {
 		i := 0
-		for yyExca[i] != -1 || yyExca[i+1] != state {
+		for yyExca[i] != -1 || int(yyExca[i+1]) != state {
 			i += 2
 		}
 
 		// Look for tokens that we accept or reduce.
 		for i += 2; yyExca[i] >= 0; i += 2 {
-			tok := yyExca[i]
+			tok := int(yyExca[i])
 			if tok < TOKSTART || yyExca[i+1] == 0 {
 				continue
 			}
@@ -720,30 +720,30 @@ func yylex1(lex yyLexer, lval *yySymType) (char, token int) {
 	token = 0
 	char = lex.Lex(lval)
 	if char <= 0 {
-		token = yyTok1[0]
+		token = int(yyTok1[0])
 		goto out
 	}
 	if char < len(yyTok1) {
-		token = yyTok1[char]
+		token = int(yyTok1[char])
 		goto out
 	}
 	if char >= yyPrivate {
 		if char < yyPrivate+len(yyTok2) {
-			token = yyTok2[char-yyPrivate]
+			token = int(yyTok2[char-yyPrivate])
 			goto out
 		}
 	}
 	for i := 0; i < len(yyTok3); i += 2 {
-		token = yyTok3[i+0]
+		token = int(yyTok3[i+0])
 		if token == char {
-			token = yyTok3[i+1]
+			token = int(yyTok3[i+1])
 			goto out
 		}
 	}
 
 out:
 	if token == 0 {
-		token = yyTok2[1] /* unknown char */
+		token = int(yyTok2[1]) /* unknown char */
 	}
 	if yyDebug >= 3 {
 		__yyfmt__.Printf("lex %s(%d)\n", yyTokname(token), uint(char))
@@ -798,7 +798,7 @@ yystack:
 	yyS[yyp].yys = yystate
 
 yynewstate:
-	yyn = yyPact[yystate]
+	yyn = int(yyPact[yystate])
 	if yyn <= yyFlag {
 		goto yydefault /* simple state */
 	}
@@ -809,8 +809,8 @@ yynewstate:
 	if yyn < 0 || yyn >= yyLast {
 		goto yydefault
 	}
-	yyn = yyAct[yyn]
-	if yyChk[yyn] == yytoken { /* valid shift */
+	yyn = int(yyAct[yyn])
+	if int(yyChk[yyn]) == yytoken { /* valid shift */
 		yyrcvr.char = -1
 		yytoken = -1
 		yyVAL = yyrcvr.lval
@@ -823,7 +823,7 @@ yynewstate:
 
 yydefault:
 	/* default state action */
-	yyn = yyDef[yystate]
+	yyn = int(yyDef[yystate])
 	if yyn == -2 {
 		if yyrcvr.char < 0 {
 			yyrcvr.char, yytoken = yylex1(yylex, &yyrcvr.lval)
@@ -832,18 +832,18 @@ yydefault:
 		/* look through exception table */
 		xi := 0
 		for {
-			if yyExca[xi+0] == -1 && yyExca[xi+1] == yystate {
+			if yyExca[xi+0] == -1 && int(yyExca[xi+1]) == yystate {
 				break
 			}
 			xi += 2
 		}
 		for xi += 2; ; xi += 2 {
-			yyn = yyExca[xi+0]
+			yyn = int(yyExca[xi+0])
 			if yyn < 0 || yyn == yytoken {
 				break
 			}
 		}
-		yyn = yyExca[xi+1]
+		yyn = int(yyExca[xi+1])
 		if yyn < 0 {
 			goto ret0
 		}
@@ -865,10 +865,10 @@ yydefault:
 
 			/* find a state where "error" is a legal shift action */
 			for yyp >= 0 {
-				yyn = yyPact[yyS[yyp].yys] + yyErrCode
+				yyn = int(yyPact[yyS[yyp].yys]) + yyErrCode
 				if yyn >= 0 && yyn < yyLast {
-					yystate = yyAct[yyn] /* simulate a shift of "error" */
-					if yyChk[yystate] == yyErrCode {
+					yystate = int(yyAct[yyn]) /* simulate a shift of "error" */
+					if int(yyChk[yystate]) == yyErrCode {
 						goto yystack
 					}
 				}
@@ -904,7 +904,7 @@ yydefault:
 	yypt := yyp
 	_ = yypt // guard against "declared and not used"
 
-	yyp -= yyR2[yyn]
+	yyp -= int(yyR2[yyn])
 	// yyp is now the index of $0. Perform the default action. Iff the
 	// reduced production is ε, $1 is possibly out of range.
 	if yyp+1 >= len(yyS) {
@@ -915,16 +915,16 @@ yydefault:
 	yyVAL = yyS[yyp+1]
 
 	/* consult goto table to find next state */
-	yyn = yyR1[yyn]
-	yyg := yyPgo[yyn]
+	yyn = int(yyR1[yyn])
+	yyg := int(yyPgo[yyn])
 	yyj := yyg + yyS[yyp].yys + 1
 
 	if yyj >= yyLast {
-		yystate = yyAct[yyg]
+		yystate = int(yyAct[yyg])
 	} else {
-		yystate = yyAct[yyj]
-		if yyChk[yystate] != -yyn {
-			yystate = yyAct[yyg]
+		yystate = int(yyAct[yyj])
+		if int(yyChk[yystate]) != -yyn {
+			yystate = int(yyAct[yyg])
 		}
 	}
 	// dummy call; replaced with literal code
